לשים מתחת ל
Code injection consists of running code in the memory of another process. Generally using
DLL injection but there are other options and even an entire EXE can be injected. The
complexity of the process lies in the fact that the injected code must nd a way to execute
itself without being loaded by the system (especially since the base address is not the same).
For DLL injection this is done when the DLL is loaded. In order to inject code, the code
should be able to change its memory pointers based on Reallocation.