לשים מתחת לחלק של הCRYPTOS
This mechanism cryptographically modies the program / activation code and per-
forms a decoding or sub-process function like stub. When the program is sent it is
encrypted and the decryption plug is hidden. Make it seem completely useless for AV,
thus bypassing it. Decrypting an encrypted code works in memory, and a decrypted
activation code remains on the disk. Hence AV could not remove it before performing
it.